+++ src/Resources/SubscriptionsResource.php
@@ -8,8 +8,11 @@ use Saloon\Exceptions\Request\FatalRequestException;
 use Saloon\Exceptions\Request\RequestException;
 use Saloon\Http\BaseResource;
 use Subster\PhpSdk\DataObjects\ChangeSubscriptionPlanData;
+use Subster\PhpSdk\DataObjects\RecordSubscriptionUsageEventData;
 use Subster\PhpSdk\DataObjects\SubscriptionPlanChangeData;
+use Subster\PhpSdk\DataObjects\SubscriptionUsageEventData;
 use Subster\PhpSdk\Requests\ChangeSubscriptionPlanRequest;
+use Subster\PhpSdk\Requests\RecordSubscriptionUsageEventRequest;
 
 class SubscriptionsResource extends BaseResource
 {
@@ -21,6 +24,17 @@ class SubscriptionsResource extends BaseResource
     {
         return $this->connector->send(
             new ChangeSubscriptionPlanRequest($subscription, $data)
+        )->dtoOrFail();
+    }
+
+    /**
+     * @throws FatalRequestException
+     * @throws RequestException
+     */
+    public function recordUsage(string $subscription, RecordSubscriptionUsageEventData $data): SubscriptionUsageEventData
+    {
+        return $this->connector->send(
+            new RecordSubscriptionUsageEventRequest($subscription, $data)
         )->dtoOrFail();
     }
 }